+++ src/main/java/net/jacksum/gui/UserInputError.java
@@ -1,6 +1,6 @@
 /*
 
-  HashGarten 0.15.0 - a GUI to calculate and verify hashes, powered by Jacksum
+  HashGarten 0.20.0 - a GUI to calculate and verify hashes, powered by Jacksum
   Copyright (c) 2022-2024 Dipl.-Inf. (FH) Johann N. Löfflmann,
   All Rights Reserved, <https://jacksum.net>.