+++ src/main/java/net/jacksum/gui/constants/AppConstants.java
@@ -1,6 +1,6 @@
 /*
 
-  HashGarten 0.19.0 - a GUI to calculate and verify hashes, powered by Jacksum
+  HashGarten 0.20.0 - a GUI to calculate and verify hashes, powered by Jacksum
   Copyright (c) 2022-2024 Dipl.-Inf. (FH) Johann N. Löfflmann,
   All Rights Reserved, <https://jacksum.net>.
 
@@ -25,7 +25,7 @@ package net.jacksum.gui.constants;
  * @author Johann
  */
 public class AppConstants {
-    public final static String VERSION = "0.19.0";
+    public final static String VERSION = "0.20.0";
     public final static String PROPERTIES_FILE = String.format("%s/.HashGarten.properties", System.getProperty("user.home"));
     public final static String TIMESTAMP_DEFAULT = "yyyyMMddHHmmssSSS";
 }